+++ assets/js/components/ChargingPlans/Warnings.vue
@@ -21,7 +21,8 @@
 <script lang="ts">
 import { defineComponent, type PropType } from "vue";
 import formatter from "../../mixins/formatter.js";
-import type { PlanWrapper, Tariff } from "assets/js/types/evcc.js";
+import type { PlanWrapper } from "./types.js";
+import type { Tariff } from "assets/js/types/evcc.js";
 
 export default defineComponent({
 	name: "ChargingPlanWarnings",
