+++ assets/js/types/evcc.ts
@@ -1,3 +1,5 @@
+import type { StaticPlan, RepeatingPlan } from "../components/ChargingPlans/types";
+
 declare global {
 	interface Window {
 		app: any;
@@ -25,35 +27,8 @@ export interface Vehicle {
 	title: string;
 }
 
-export interface PlanWrapper {
-	planId: number;
-	planTime: Date;
-	duration: number;
-	plan: Rate[];
-	power: number;
-}
-
-export type StaticPlan = StaticSocPlan | StaticEnergyPlan;
-
-export interface StaticSocPlan {
-	soc: number;
-	time: Date;
-}
-
-export interface StaticEnergyPlan {
-	energy: number;
-	time: Date;
-}
-
 export type Timeout = ReturnType<typeof setInterval> | null;
 
-export interface RepeatingPlan {
-	weekdays: number[];
-	time: string;
-	tz: string; // timezone like "Europe/Berlin"
-	soc: number;
-	active: boolean;
-}
 export interface Tariff {
 	rates: Rate[];
 	lastUpdate: Date;
